+++ NEWS.md
@@ -1,6 +1,10 @@
 
 # renv (development version)
 
+* Fixed an issue where `renv` could stall when loading a project and validating
+  that the configured version of Bioconductor is compatible with the version
+  of R currently in use.
+
 * `renv::snapshot()` no longer fails to generate a new lockfile if the project
   contains a lockfile which could not be read or parsed.